+++ packages/terra-brand-footer/src/terra-dev-site/test/brand-footer/DefaultBrandFooter.test.jsx
@@ -90,4 +90,5 @@ export default () => (
         </span>
 )}
     />
-  </div>);
+  </div>
+);