+++ packages/terra-brand-footer/src/terra-dev-site/test/brand-footer/LinksPropBrandFooter.test.jsx
@@ -85,4 +85,5 @@ export default () => (
         </span>
 )}
     />
-  </div>);
+  </div>
+);